Extract click position mapper for input fields for re-use
