+++ spec/features/work_packages/table/context_menu_spec.rb
@@ -97,7 +97,7 @@ describe 'Work package table context menu', js: true do
       wp_table.expect_work_package_listed(work_package2)
 
       # Select both
-      all('td.checkbox input').each { |el| el.set(true) }
+      wp_table.table_container.send_keys [:control, 'a']
     end
 
     it 'shows a subset of the available menu items' do
